scripts/requirements.txt
# Python dependencies for Pyannote speaker diarization
# Install with: pip install -r requirements.txt

pyannote.audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# Note: torch and torchaudio will download ~2GB on first install
# For CPU-only (lighter):
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
